home/apps/difflens/pyproject.toml
[build-system]
requires = ["setuptools==69.5.1", "wheel==0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "difflens"
version = "0.1.0"
description = "Deterministic text and JSON diff service."
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "psycopg[binary]==3.2.13",
  "pydantic==2.10.6",
  "python-dotenv==1.0.1"
]

[tool.setuptools]
packages = ["app"]
